2021-01-11 (지산) 씬 뒤로가기 추가 (backspace - 이미지 추가 고민중) 볼륨조절 width 수정 볼륨조절화살표, 캐릭터선택화살표 배경 추가 UPslope, DOWNslope rect비교 ptrc -> shdrc

+++ settingScene.h
@@ -5,8 +5,8 @@ class settingScene : public gameNode
 {
 private:
 	//volume width ·Î render ÁöÁ¤
-	int _background_volume_width;
-	int _SFX_volume_width;
+	float _background_volume_width;
+	float _SFX_volume_width;
 	
 	image* _background_volume;
 	image* _SFX_volume;
